requirements.txt
anyio==3.7.1
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
exceptiongroup==1.2.0
idna==3.6
ombott==1.0.0
pluralize==20231008.1
portalocker==2.8.2
py4web==1.20231115.1
pycparser==2.21
pycryptodome==3.19.0
pydal==20231114.3
PyJWT==2.8.0
renoir==1.7.0
requests==2.31.0
rocket3==20231022.2
sniffio==1.3.0
threadsafevariable==20230507.1
tornado==6.4
urllib3==2.1.0
watchgod==0.8.2
yatl==20230507.3
python-socketio==5.11.0
greenlet==3.0.3
gevent==23.9.1
simple-websocket==1.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pytest==7.4.3
redis==5.0.1
google-generativeai==0.3.2
